requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Pyrebase4==4.3.0
python-firebase==1.2
requests==2.22.0
joblib

